Fix for keyboard I/O

+++ SMARTApi.pas
@@ -49,8 +49,8 @@ var
   SMART_MouseBtnUp: procedure(const Target: PTarget; const X, Y: Integer; const Btn: Integer); cdecl;
   SMART_GetMouseBtnState: function(const Target: PTarget; const Btn: Integer): Boolean; cdecl;
   SMART_TypeText: procedure(const Target: PTarget; const Str: PAnsiChar; const KeyWait, KeyModWait: Integer); cdecl;
-  SMART_VKeyDown: procedure(const Target: PTarget; const Key: Byte); cdecl;
-  SMART_VKeyUp: procedure(const Target: PTarget; const Key: Byte); cdecl;
+  SMART_VKeyDown: procedure(const Target: PTarget; const Key: Integer); cdecl;
+  SMART_VKeyUp: procedure(const Target: PTarget; const Key: Integer); cdecl;
   SMART_GetKeyState: function(const Target: PTarget; const Key: Byte): Boolean; cdecl;
   SMART_ClientID: function(const Idx: Integer): Integer; cdecl;
   SMART_GetClients: function(const OnlyUnpaired: Boolean): Integer; cdecl;
